refacto create trait instead of function in AbstractBuilder
