+++ docs/pdf/ConvertPdfBuilder.md
@@ -72,12 +72,12 @@ return $gotenberg
 ```
 
 ### files(Stringable|string ...\$paths)
-Adds files for the builder (overrides any previous files).<br />
+Adds files (overrides any previous files).<br />
 
 ```php
 return $gotenberg
     // Your builder call as ->html() and the rest of your configuration code
-    ->files('document.pdf', __DIR__'/../../public/document_2.pdf')
+    ->files('document.pdf', '/absolute/path/document_2.pdf')
     ->generate()
     ->stream()
 ;
