+++ src/Builder/AbstractBuilder.php
@@ -10,7 +10,6 @@ use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\LoggerAwareTrait;
 use Sensiolabs\GotenbergBundle\Builder\Result\GotenbergAsyncResult;
 use Sensiolabs\GotenbergBundle\Builder\Result\GotenbergFileResult;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
-use Sensiolabs\GotenbergBundle\Builder\Util\ValidatorFactory;
 use Sensiolabs\GotenbergBundle\Client\GotenbergClientInterface;
 use Sensiolabs\GotenbergBundle\Exception\InvalidNormalizerException;
 use Sensiolabs\GotenbergBundle\Exception\VersionCompatibilityException;
@@ -49,11 +48,6 @@ abstract class AbstractBuilder implements BuilderAsyncInterface, BuilderFileInte
     }
 
     abstract protected function getEndpoint(): string;
-
-    protected function getAllowedFilesExtensions(): array
-    {
-        return [];
-    }
 
     /**
      *  The API automatically appends the file extension, so there's no need for you to set it manually.
@@ -199,25 +193,5 @@ abstract class AbstractBuilder implements BuilderAsyncInterface, BuilderFileInte
 
             yield from $normalizer($key, $value, $version, $logger);
         }
-    }
-
-    /**
-     * Adds files (overrides any previous files).
-     *
-     * @example files('document.pdf', '/absolute/path/document_2.pdf')
-     */
-    public function files(string|\Stringable ...$paths): self
-    {
-        foreach ($paths as $path) {
-            $path = (string) $path;
-            $info = new \SplFileInfo($this->getAssetBaseDirFormatter()->resolve($path));
-            ValidatorFactory::filesExtension([$info], $this->getAllowedFilesExtensions());
-
-            $files[$path] = $info;
-        }
-
-        $this->getBodyBag()->set('files', $files ?? null);
-
-        return $this;
     }
 }
