+++ src/Builder/Pdf/LibreOfficePdfBuilder.php
@@ -8,6 +8,7 @@ use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\AssetBaseDirFormatterAwareTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\EmbedTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\EncryptTrait;
+use Sensiolabs\GotenbergBundle\Builder\Behaviors\FilesTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\LibreOfficeTrait;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
 use Sensiolabs\GotenbergBundle\Enumeration\SplitMode;
@@ -29,6 +30,7 @@ final class LibreOfficePdfBuilder extends AbstractBuilder
     use AssetBaseDirFormatterAwareTrait;
     use EmbedTrait;
     use EncryptTrait;
+    use FilesTrait;
     use LibreOfficeTrait;
 
     public const ENDPOINT = '/forms/libreoffice/convert';
