+++ src/Builder/Screenshot/MarkdownScreenshotBuilder.php
@@ -6,6 +6,7 @@ use Sensiolabs\GotenbergBundle\Builder\AbstractBuilder;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\NormalizeGotenbergPayload;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\ChromiumScreenshotTrait;
+use Sensiolabs\GotenbergBundle\Builder\Behaviors\FilesTrait;
 use Sensiolabs\GotenbergBundle\Builder\BuilderAssetInterface;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
 use Sensiolabs\GotenbergBundle\Enumeration\Part;
@@ -26,6 +27,7 @@ final class MarkdownScreenshotBuilder extends AbstractBuilder implements Builder
         content as private;
         contentFile as private;
     }
+    use FilesTrait;
 
     public const ENDPOINT = '/forms/chromium/screenshot/markdown';
